requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
easyocr==1.7.0
filelock==3.12.1
fonttools==4.39.4
gunicorn==20.1.0
idna==3.4
imageio==2.31.0
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
ninja==1.11.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
psutil==5.9.5
pyclipper==1.3.0.post4
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sympy==1.12
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.116
urllib3==2.0.3
